Divisão do header/index em server e client
